backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
httpx>=0.26.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-socketio[asyncio]>=5.10.0
